Add fixes to overhead_window_from_events and to test_validator.py
