+++ tests/validator/test_overhead_from_events.py
@@ -1,7 +1,4 @@
-from dataclasses import dataclass
 from datetime import datetime, timedelta
-from enum import Enum, auto
-from typing import List
 
 from satellite_determination.custom_dataclasses.coordinates import Coordinates
 from satellite_determination.custom_dataclasses.facility import Facility
@@ -9,60 +6,8 @@ from satellite_determination.custom_dataclasses.overhead_window import OverheadW
 from satellite_determination.custom_dataclasses.reservation import Reservation
 from satellite_determination.custom_dataclasses.satellite.satellite import Satellite
 from satellite_determination.custom_dataclasses.time_window import TimeWindow
-
-
-class EventTypesRhodesmill(Enum):
-    ENTERS = 0 #Skyfield API returns events as 0, 1, or 2 for enters, culminates, exits
-    CULMINATES = 1
-    EXITS = 2
-
-
-@dataclass
-class EventRhodesmill:
-    event_type: EventTypesRhodesmill
-    satellite: Satellite
-    timestamp: datetime
-
-
-class OverheadWindowFromEvents:
-    def __init__(self, events: List[EventRhodesmill], reservation: Reservation):
-        self._events = events
-        self._reservation = reservation
-
-    def get(self) -> List[OverheadWindow]:
-        enter_events, culminate_events, exit_events = ([event for event in self._events if event.event_type == event_type]
-                                     for event_type in EventTypesRhodesmill)
-        print(len(enter_events))
-        print(len(culminate_events))
-        print(len(exit_events))
-
-        if (len(self._events) == 1) and (self._events[0].event_type == EventTypesRhodesmill.CULMINATES): #handles sat that is in view for entire reservation
-            time_window = TimeWindow(begin=self._reservation.time.begin, end=self._reservation.time.end)
-            overhead_windows = [OverheadWindow(satellite=self._events[0].satellite, overhead_time=time_window)]
-        #ADD TEST CASE FOR: 1 2 0 1 CASE 1201
-        else:
-            if len(enter_events) != len(exit_events): #Handle case where a satellite starts or ends in observation area
-                if self._events[0].event_type == EventTypesRhodesmill.ENTERS: #the first event is the satellite entering view, so it didn't start in observation area
-                    end_reservation_event = EventRhodesmill(event_type=EventTypesRhodesmill.EXITS, satellite=self._events[0].satellite, timestamp=self._reservation.time.end)
-                    exit_events.append(end_reservation_event)
-                elif self._events[0].event_type == EventTypesRhodesmill.EXITS or self._events[0].event_type == EventTypesRhodesmill.CULMINATES: #the first event is an exit, so the sat starts in view
-                    start_reservation_event = EventRhodesmill(event_type=EventTypesRhodesmill.ENTERS, satellite=self._events[0].satellite, timestamp=self._reservation.time.begin)
-                    enter_events.insert(0, start_reservation_event)
-            elif len(enter_events) == len(exit_events) and (self._events[0].event_type == EventTypesRhodesmill.EXITS or self._events[0].event_type == EventTypesRhodesmill.CULMINATES):
-                if self._events[0] == 1:
-                    start_reservation_event = EventRhodesmill(event_type=EventTypesRhodesmill.ENTERS, satellite=self._events[0].satellite, timestamp=self._reservation.time.begin)
-                    enter_events.insert(0, start_reservation_event)
-                    end_reservation_event = EventRhodesmill(event_type=EventTypesRhodesmill.EXITS, satellite=self._events[0].satellite, timestamp=self._reservation.time.end)
-                    exit_events.append(end_reservation_event)
-                else:
-                    start_reservation_event = EventRhodesmill(event_type=EventTypesRhodesmill.ENTERS, satellite=self._events[0].satellite, timestamp=self._reservation.time.begin)
-                    enter_events.insert(0, start_reservation_event)
-                    end_reservation_event = EventRhodesmill(event_type=EventTypesRhodesmill.EXITS, satellite=self._events[0].satellite, timestamp=self._reservation.time.end)
-                    exit_events.append(end_reservation_event)
-            enter_and_exit_pairs = zip(enter_events, exit_events)
-            time_windows = [TimeWindow(begin=begin_event.timestamp, end=exit_event.timestamp) for begin_event, exit_event in enter_and_exit_pairs]
-            overhead_windows = [OverheadWindow(satellite=self._events[0].satellite, overhead_time=time_window) for time_window in time_windows]
-        return overhead_windows
+from satellite_determination.validator.validator_skyfield.support.overhead_window_from_events import \
+    EventTypesRhodesmill, EventRhodesmill, OverheadWindowFromEvents
 
 
 class TestOverheadWindowFromEvents:
@@ -391,6 +336,48 @@ class TestOverheadWindowFromEvents:
                 )
             )
         ]
+
+    def test_satellite_0_1_2_0(self):
+        events = [
+            EventRhodesmill(
+                event_type=EventTypesRhodesmill.ENTERS,
+                satellite=self._arbitrary_satellite,
+                timestamp=self._arbitrary_date
+            ),
+            EventRhodesmill(
+                event_type=EventTypesRhodesmill.CULMINATES,
+                satellite=self._arbitrary_satellite,
+                timestamp=self._arbitrary_date + timedelta(hours=1)
+            ),
+            EventRhodesmill(
+                event_type=EventTypesRhodesmill.EXITS,
+                satellite=self._arbitrary_satellite,
+                timestamp=self._arbitrary_date + timedelta(hours=2)
+            ),
+            EventRhodesmill(
+                event_type=EventTypesRhodesmill.ENTERS,
+                satellite=self._arbitrary_satellite,
+                timestamp=self._arbitrary_date + timedelta(hours=3)
+            ),
+        ]
+        overhead_windows = OverheadWindowFromEvents(events=events, reservation=self._arbitrary_reservation_with_nonzero_timewindow).get()
+        assert overhead_windows == [
+            OverheadWindow(
+                satellite=self._arbitrary_satellite,
+                overhead_time=TimeWindow(
+                    begin=events[0].timestamp,
+                    end=events[2].timestamp
+                )
+            ),
+            OverheadWindow(
+                satellite=self._arbitrary_satellite,
+                overhead_time=TimeWindow(
+                    begin=events[3].timestamp,
+                    end=self._arbitrary_reservation_with_nonzero_timewindow.time.end
+                )
+            )
+        ]
+
 #Could a sat be in view but not trigger any of the events?
     @property
     def _arbitrary_satellite(self) -> Satellite:
