+++ tests/validator/test_validator.py
@@ -3,19 +3,17 @@ import datetime
 import os
 from pathlib import Path
 from datetime import datetime
-import json
 import filecmp
 
 from satellite_determination.custom_dataclasses.coordinates import Coordinates
 from satellite_determination.custom_dataclasses.facility import Facility
 from satellite_determination.custom_dataclasses.reservation import Reservation
-#from satellite_determination.custom_dataclasses.overhead_window import OverheadWindow
 from satellite_determination.retrievers.satellite_retriever.skyfield_satellite_retriever import SkyfieldSatelliteList
-#from satellite_determination.validator.validator import Validator
 from satellite_determination.custom_dataclasses.time_window import TimeWindow
 from satellite_determination.utilities import convert_dt_to_utc
 from tests.utilities import get_script_directory
-from tests.validator.test_overhead_from_events import EventRhodesmill, OverheadWindowFromEvents, EventTypesRhodesmill
+from satellite_determination.validator.validator_skyfield.support.overhead_window_from_events import \
+    EventTypesRhodesmill, EventRhodesmill, OverheadWindowFromEvents
 
 
 class ValidatorRhodesMill:
@@ -31,13 +29,10 @@ class ValidatorRhodesMill:
         t1 = ts.utc(convert_dt_to_utc(self._reservation.time.end))
         coordinates = wgs84.latlon(self._reservation.facility.point_coordinates.latitude, self._reservation.facility.point_coordinates.longitude)
         for sat in self._list_of_satellites.satellites:
-            print(sat.name)
             t, events = sat.find_events(coordinates, t0, t1, altitude_degrees=self._reservation.facility.angle_of_visibility_cone)
             if events.size == 0:
-                print("no events")
                 continue
             else:
-                #skeleton
                 rhodesmill_event_list = []
                 for ti, event in zip(t, events):
                     #print(event)
@@ -53,11 +48,6 @@ class ValidatorRhodesMill:
                 sat_windows = OverheadWindowFromEvents(events=rhodesmill_event_list, reservation=self._reservation).get()
                 for window in sat_windows:
                     overhead_windows.append(window)
-                if len(sat_windows) != 0:
-                    print(sat_windows)
-                else:
-                    print("none")
-        #print(overhead_windows[0].overhead_time)
         return overhead_windows
